Fixing bug reported and patched by Christian Frichot.
"""
The issue appeared to stem back to the incorrect call of
addToHighlight() method within plugins/grep/codeDisclosure.py. The
"res" output from the grep call is an object, to get the text from the
object you need to call "res.group()"
"""

This was reported to the w3af-develop mailing list.