Update update.rst
Line 7, added the word version to clarify what is being updated. The w3af version is being updated.
Line 16, added comma after update for grammar correction.  
Line 44, added quotes around the last ''feature'' entry to make it highlighted as the other instances of the branches. removed 'which' from ...disturb other users which require... and replace with 'that' for grammar.