Update phases.rst
Line 4, removed "the basis about" to clarify statement.
Line 14, removed 'are' and 'knowledge' and added clarifying words 'have' and 'information respectively to clarify language.
Line 24, added 'of' in the statement ...or a dump of remote tables... to correct for grammar. Removed the 's' from injections for grammar.
Lines 45 and 46, added the '' for Enable all to match formatting.
Line 57, added period and created new sentence in order to avoid run-on. Changed 'on' to 'in' and removed 'inside' to keep meaning of sentence clear.