Added test for multiple requests, same cookie
Added unit test for the scenario where the same secure cookie
is sent over HTTP to multiple paths. Currently the test fails,
and the problem lies in how the vulnerability is being submitted
to the database. Attempting to do a lookup and compare of existing
vulns in the database fails since the URI comparison returns false
due to the different paths being included in the URI.