Covered test case of same vuln over same url
In the case of an updated cookie, I added a test case for when a cookie
is submitted over the same url with the same vulnerability it will
merely update the url. Also when update runs now, if no vulnerability of
the same name is found, then the function will append the vulnerability
as a new entry.