This commit fixes an issue introduced when we added more payloads to the global_redirect plugin: before we used 2 HTTP requests per parameter, after the improvement we used ~10.
More requests mean more coverage, and more chances of identifying a vulnerability. But also mean long scan times.

I changed a couple of things in the plugin in order to decide when it makes sense to send all the "extended payload set". Also added more tests to make sure everything is working.