Refactored very aggressive lock which was blocking numerous threads for the time required to perform 3 HTTP requests.
This doesn't sound like a lot, but when the servers answers requests in 1 second, this is blocking blind sqli, os commanding, eval, and any other get_average_rtt_for_mutant() user for 3 seconds.
If there are many calls to get_average_rtt_for_mutant() this will delay the scan for several minutes.